backend/app/routers/Project3/requirements.txt
streamlit
tensorflow
numpy
Pillow
matplotlib
scikit-learn